remove the need of default query generator
