+++ public/app/features/trails/autoQuery/types.ts
@@ -17,3 +17,10 @@ export interface AutoQueryInfo {
 }
 
 export type VizBuilder = () => VizPanelBuilder<{}, {}>;
+
+export type AutoQueryContext = {
+  metricParts: string[];
+  isRateQuery: boolean;
+  isUtf8Metric: boolean;
+  baseQuery: string;
+};
